requirements.txt
# Core scientific stack
numpy>=1.26
scipy>=1.12
pandas>=2.2
matplotlib>=3.8
seaborn>=0.13
PyYAML>=6.0
tabulate>=0.9
tqdm>=4.66

# RL / simulation stack
gymnasium>=0.29
torch-pso>=1.2.1
stable-baselines3>=2.8.0
sb3-contrib>=2.8.0
shimmy>=1.3

# Notebook / reporting
jupyterlab>=4.1
notebook>=7.1

# Testing
pytest>=8.1
